requirements.txt
# Python 2.6 compatibility
# argparse==1.2.1
pygithub==1.37
Flask==0.12.2
